video: common: addition of a first skeleton for helpers
This is a trial of a first set of helpers for the video subsystem
in order to handling queue and m2m devices.  It is to be enhanced
if appropriate, we can consider that some more elements could be
added into the video structure such as format, lock, since most
of device have (or should have), allowing to enforce some of the
API behaviors of the video subsystems (such as, not changing the
fmt while streaming, stream_start only once queued etc)

DISCLAIMER: this is simply done for sharing ideas, there is a very
good chance that this does not even compile properly yet.

Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>